+++ base/src/main/java/io/spine/reflect/Types.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.base;
+package io.spine.reflect;
 
 import com.google.common.reflect.TypeParameter;
 import com.google.common.reflect.TypeToken;
